requirements.txt
Django==1.10
django-bootstrap-pagination==1.6.2
django-environ==0.4.0
django-extensions==1.7.1
django-model-utils==2.5.2
djangorestframework==3.4.4
gunicorn==19.6.0
ipython==5.1.0
Markdown==2.6.7
psycopg2==2.6.2
PyGithub==1.29
pytz==2016.10
six==1.10.0
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
whitenoise==3.2.1
